requirements.txt
asyncio==3.4.3
attrs==24.2.0
fastavro==1.9.7
graphql-core==3.2.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
memphis-py==1.3.1
mmh3==5.0.1
nats-py==2.9.0
paho-mqtt==2.1.0
protobuf==5.28.2
referencing==0.35.1
rpds-py==0.20.0